Improve reporting for key values
With path normalization (and even before path normalization), it was
possible for the reported path to not match the input value. For
example, with environment variables, the environment variable may have
been `FOO__BAR`, but the report showed `FOO.BAR`, which makes it harder
for users to trace the report back to the original source.

We now add a `sourceKey` attribute to `Node` and report on it, to show
user's the key value as it was originally present in the source.