Enable path normalization by default
Remove both the `SnakeCaseParamMapper` and `KebabCaseParamMapper`
by default, add the `PathNormalizer` by default.

Add removal of `_` to path normalizer.

Fix some issues with the `HikariDataSourceDecoder` when enabling
path normalization by default -- that decoder requires the
original key case as its props are case-sensitive. Create an
abstract `UnnormalizedKeysDecoder` which has the ability to restore
the case of keys via the `sourceKey`.

Fix breaking explicit sealed types with normalization because the
discriminator field defaults to `_type` which normalizes to `type`.
Disable normalization if the field name matches the discriminator field
name, and the node is a `MapNode`.

Fix reporting for strict mode to use the `sourceKey` value, so that
reporting matches the source value, not the normalized value.

Update Preprocessor implementations to correctly copy the source key
when new Map and Array nodes are created.