Data class decoder can use node transformers
When the data class decoder is decoding, we may have @ConfigAlias
names. However, these names are not transformed using the node
transformers, and therefore may not match the nodes.

We now add the configured node transformers to the decoding context
so decoders have access to these.

The `DataClassDecoder` uses the transformers to convert the
possible param names, so we are comparing apples-to-apples.